♻️(api): move modulix script
in devcomp context
