--- api/scripts/modulix/dev-convert-joi-to-json-schema.js
@@ -1,6 +1,0 @@
-import { convertJoiToJsonSchema } from '../../src/devcomp/infrastructure/datasources/conversion/joi-to-json-schema.js';
-import { moduleSchema } from '../../src/devcomp/infrastructure/datasources/learning-content/validation/module-schema.js';
-
-const jsonSchema = JSON.stringify(convertJoiToJsonSchema(moduleSchema), null, 2);
-
-console.log('JSON schema generated by convertJoiToJsonSchema :', jsonSchema);
