+++ api/src/devcomp/scripts/dev-convert-joi-to-json-schema.js
@@ -1,0 +1,9 @@
+// eslint-disable-next-line @eslint-community/eslint-comments/disable-enable-pair
+/* eslint-disable no-console */
+
+import { convertJoiToJsonSchema } from '../infrastructure/datasources/conversion/joi-to-json-schema.js';
+import { moduleSchema } from '../infrastructure/datasources/learning-content/validation/module-schema.js';
+
+const jsonSchema = JSON.stringify(convertJoiToJsonSchema(moduleSchema), null, 2);
+
+console.log('JSON schema generated by convertJoiToJsonSchema :', jsonSchema);
