+++ api/src/devcomp/scripts/get-modules-assets-debt.js
@@ -1,4 +1,7 @@
-import moduleDatasource from '../../src/devcomp/infrastructure/datasources/learning-content/module-datasource.js';
+// eslint-disable-next-line @eslint-community/eslint-comments/disable-enable-pair
+/* eslint-disable no-console */
+
+import moduleDatasource from '../infrastructure/datasources/learning-content/module-datasource.js';
 import { getElements } from './get-elements-csv.js';
 
 async function main() {
