+++ api/src/devcomp/scripts/utils/get-answerable-elements.js
