+++ api/tests/devcomp/acceptance/scripts/get-elements-csv_test.js
@@ -1,8 +1,8 @@
 import nock from 'nock';
 import sinon from 'sinon';
 
-import { getElements, getElementsListAsCsv } from '../../../../scripts/modulix/get-elements-csv.js';
 import { ModuleFactory } from '../../../../src/devcomp/infrastructure/factories/module-factory.js';
+import { getElements, getElementsListAsCsv } from '../../../../src/devcomp/scripts/get-elements-csv.js';
 import { expect } from '../../../test-helper.js';
 import moduleContent from './test-module.json' with { type: 'json' };
