+++ api/tests/devcomp/acceptance/scripts/get-sections-csv_test.js
@@ -1,7 +1,7 @@
 import nock from 'nock';
 
-import { getSectionsListAsCsv } from '../../../../scripts/modulix/get-sections-csv.js';
 import { ModuleFactory } from '../../../../src/devcomp/infrastructure/factories/module-factory.js';
+import { getSectionsListAsCsv } from '../../../../src/devcomp/scripts/get-sections-csv.js';
 import { expect } from '../../../test-helper.js';
 import moduleContent from './test-module.json' with { type: 'json' };
