+++ api/tests/devcomp/acceptance/scripts/utils/get-grains_test.js
@@ -1,4 +1,4 @@
-import { getGrains } from '../../../../../scripts/modulix/utils/get-grains.js';
+import { getGrains } from '../../../../../src/devcomp/scripts/utils/get-grains.js';
 import { expect } from '../../../../test-helper.js';
 
 describe('Acceptance | Script | Helper | Get grains', function () {
